moving all TeiInteract plugin files to the root level
Previously, this repository had been rooted at the Omeka application level.
Since the focus of development is on this plugin, it is apporpriate that it should
occupy its own repo. This TeiInteract repository is a submodule of the literati/Omeka
repository, forked from omeka/omeka and housed under literati so that any changes to core we need to
make can be done in isolation from omeka/omeka...